requirements.txt
psycopg2==2.9.3
Django==4.0.10
django-admin-interface==0.19.1
hypercorn==0.13.2
Xlsxwriter==3.0.2
whitenoise==5.3.0
reportlab==3.6.6
matplotlib==3.5.1
django-leaflet==0.28.2
django-cleanup==5.2.0
django-geojson==3.2.0
django-smart-selects==1.6.0
djangorestframework==3.13.1
djangorestframework-gis==0.18
svglib==1.2.1
